Changed how memory is allocated on Linux after initialisation, which also meant getting rid of new/delete. Removed unnecessary references on cpu_state_handle and thread_state_handle. Updated the bounds_checker policy to only do one bounds check on a SOURCE_DEST operand. Updated the cfg policy to try to build both inter- and intra-procedural control-flow graphs, although I'm not convinced that it's fully working. In user space, cfg segfaults (potentially because I'm not saving/restoring the xmm registers, that is a TODO), and in kernel space, it seems like both inter- and intra-procedural graphs are overly simplistic.

+++ deps/dr/globals.h
@@ -355,7 +355,7 @@ extern "C" {
 
 extern void *heap_alloc(void *, unsigned long long);
 extern void *granary_heap_alloc_temp_instr(void);
-extern void heap_free(void *, void *, unsigned long long);
+extern void heap_free(void *, void *, unsigned long);
 extern dcontext_t *get_thread_private_dcontext(void);
 
 extern void granary_break_on_fault(void);
